chore: add classname property to button
