+++ js/react/lib/components/button/button.component.tsx
@@ -1,51 +1,24 @@
 import { ReactNode } from "react";
 import { cn } from "../../utils/tw-merge";
-
-export const variants = {
-  primary: [
-    "border rounded-2xl",
-    "bg-primary-500 border-black text-black shadow-rb-black",
-    "dark:bg-primary-300",
-    "hover:bg-primary-600 dark:hover:bg-primary-400 active:shadow-none",
-    "disabled:bg-neutral-100 disabled:shadow-none disabled:border-neutral-400 disabled:text-neutral-400",
-    "dark:disabled:bg-neutral-950",
-  ],
-  secondary: [
-    "border rounded-2xl",
-    "bg-light text-neutral-950 border-neutral-950 shadow-rb-neutral-950",
-    "dark:bg-dark dark:text-light dark:border-light dark:shadow-rb-neutral-50",
-    "hover:shadow-rb-primary-500 hover:border-primary-500 hover:text-primary-500",
-    "disabled:bg-neutral-100 disabled:text-neutral-400 disabled:border-neutral-400! disabled:shadow-none",
-    "dark:disabled:bg-neutral-950",
-    "active:shadow-none",
-  ],
-  text: [
-    "text-neutral-950 hover:text-primary-600",
-    "dark:text-light dark:hover:text-primary-300",
-  ],
-  icon: [
-    "rounded-full border aspect-square p-2! !h-fit",
-    "bg-light border-black text-black",
-    "dark:bg-dark dark:border-light dark:text-light",
-    "hover:text-primary-500 hover:border-primary-500",
-  ],
-};
+import { ButtonVariants, variants } from "./button.const";
 
 export type ButtonProps = {
-  variant?: keyof typeof variants;
+  variant?: ButtonVariants;
   label?: string;
   disabled?: boolean;
   icon?: ReactNode;
+  className?: string;
 };
 
 export const Button = (props: ButtonProps) => {
-  const { label, variant = "primary", icon, ...attr } = props;
+  const { label, variant = "primary", icon, className, ...attr } = props;
   return (
     <button
       className={cn(
         ...variants[variant],
         "px-8 h-12 text-button w-fit cursor-pointer disabled:cursor-not-allowed flex items-center justify-center gap-2.5 transition",
-        "[&>svg]:size-6"
+        "[&>svg]:size-6",
+        className
       )}
       {...attr}
     >
